+++ openlane/flows/classic.py
@@ -32,18 +32,23 @@ from .sequential import SequentialFlow
 @Flow.factory.register()
 class Classic(SequentialFlow):
     """
-    The flow most similar to the original Tcl-based OpenLane.
+    A flow of type :class:`openlane.flows.SequentialFlow` that is the most
+    similar to the original OpenLane 1.0 flow, running the Verilog RTL through
+    Yosys, OpenROAD, KLayout and Magic to produce a valid GDSII for simpler designs.
+
+    This is the default when using OpenLane via the command-line.
     """
 
     Steps: List[Type[Step]] = [
+        Yosys.JsonHeader,
         Yosys.Synthesis,
         Checker.YosysUnmappedCells,
         Misc.LoadBaseSDC,
         OpenROAD.NetlistSTA,
         OpenROAD.Floorplan,
+        Odb.SetPowerConnections,
         Odb.ManualMacroPlacement,
-        Odb.ApplyDEFTemplate,
-        OpenROAD.TapDecapInsertion,
+        OpenROAD.TapEndcapInsertion,
         OpenROAD.IOPlacement,
         Odb.CustomIOPlacement,
         OpenROAD.GeneratePDN,
@@ -58,6 +63,8 @@ class Classic(SequentialFlow):
         OpenROAD.ResizerTimingPostGRT,
         OpenROAD.DetailedRouting,
         Checker.TrDRC,
+        Odb.ReportDisconnectedPins,
+        Checker.DisconnectedPins,
         Odb.ReportWireLength,
         Checker.WireLength,
         OpenROAD.FillInsertion,