+++ openlane/scripts/odbpy/reader.py
@@ -35,6 +35,8 @@ sys.path.insert(0, os.path.dirname(os.environ["OPENLANE_ROOT"]))
 # -- END
 
 import click
+import rich
+from rich.table import Table
 from openlane.state.design_format import DesignFormat, DesignFormatByID
 
 click  # Re-export now that the environment actually works properly