+++ openlane/steps/yosys.py
@@ -27,7 +27,9 @@ from ..config import Path, Variable, StringEnum
 
 class YosysStep(TclStep):
     def get_command(self) -> List[str]:
-        return ["yosys", "-c", self.get_script_path()]
+        script_path = self.get_script_path()
+        assert isinstance(script_path, str)
+        return ["yosys", "-c", script_path]
 
     @abstractmethod
     def get_script_path(self):
@@ -35,7 +37,61 @@ class YosysStep(TclStep):
 
 
 @Step.factory.register()
+class JsonHeader(YosysStep):
+    id = "Yosys.JsonHeader"
+    inputs = []
+    outputs = [DesignFormat.JSON_HEADER]
+
+    def get_script_path(self):
+        return os.path.join(get_script_dir(), "yosys", "json_header.tcl")
+
+    config_vars = [
+        Variable(
+            "SYNTH_POWER_DEFINE",
+            Optional[str],
+            "Specifies the name of the define used to guard power and ground connections",
+            deprecated_names=["SYNTH_USE_PG_PINS_DEFINES"],
+        ),
+        Variable(
+            "SYNTH_DEFINES",
+            Optional[List[str]],
+            "Synthesis defines",
+        ),
+        Variable(
+            "SYNTH_READ_BLACKBOX_LIB",
+            bool,
+            "A flag that enable reading the full (untrimmed) liberty file as a blackbox for synthesis. Please note that this is not used in technology mapping. This should only be used when trying to preserve gate instances in the rtl of the design.",
+            default=False,
+        ),
+        Variable(
+            "VERILOG_FILES",
+            List[Path],
+            "The paths of the design's Verilog files.",
+        ),
+        Variable(
+            "VERILOG_INCLUDE_DIRS",
+            Optional[List[str]],
+            "Specifies the Verilog `include` directories.",
+        ),
+    ]
+
+
+@Step.factory.register()
 class Synthesis(YosysStep):
+    """
+    Performs synthesis and technology mapping using Yosys and ABC, emitting a
+    netlist. Requires Yosys 0.26 or higher.
+
+    Some metrics will also be extracted and updated, namely:
+
+    * ``design__instance__count``
+    * ``design__instance_unmapped__count``
+
+    If using Yosys 0.27 or higher:
+
+    * ``design__instance__area`` is also updated.
+    """
+
     id = "Yosys.Synthesis"
     inputs = []  # The input RTL is part of the configuration
     outputs = [DesignFormat.NETLIST]
@@ -79,6 +135,11 @@ class Synthesis(YosysStep):
             default="AREA 0",
         ),
         Variable(
+            "SYNTH_DEFINES",
+            Optional[List[str]],
+            "Synthesis defines",
+        ),
+        Variable(
             "SYNTH_BUFFERING",
             bool,
             "Enables `abc` cell buffering.",
@@ -93,7 +154,7 @@ class Synthesis(YosysStep):
         Variable(
             "SYNTH_READ_BLACKBOX_LIB",
             bool,
-            "A flag that enable reading the full (untrimmed) liberty file as a blackbox for synthesis. Please note that this is not used in technology mapping. This should only be used when trying to preserve gate instances in the rtl of the design.",
+            "Additionally read the liberty file(s) as a blackbox. This will allow RTL designs to incorporate explicitly declared standard cells that will not be tech-mapped or reinterpreted.",
             default=False,
         ),
         Variable(